+++ test/services/email_domain_blacklist.js
@@ -1,6 +1,8 @@
 const EmailDomainBlacklistService = require('../../services/email_domain_blacklist');
 const chai = require('../chai');
 const { deleteAllRows } = require('../db');
+const { USE_DYNAMODB } = require('../../features');
+const initLocalDynamo = require('../local_dynamo');
 
 const { expect } = chai;
 
@@ -8,9 +10,25 @@ const DOMAIN_NAME = 'near.org';
 const STALE_AT = (new Date()).toString();
 
 describe('EmailDomainBlacklistService', function () {
+    let terminateLocalDynamo;
+    before(async function() {
+        if (USE_DYNAMODB) {
+            this.timeout(10000);
+            ({ terminateLocalDynamo } = await initLocalDynamo());
+        }
+    });
+
+    after(async function() {
+        if (USE_DYNAMODB) {
+            await terminateLocalDynamo();
+        }
+    });
+
     describe('getDomainBlacklistEntry', function () {
         before(async function() {
-            await deleteAllRows();
+            if (!USE_DYNAMODB) {
+                await deleteAllRows();
+            }
         });
 
         it('gets the specified blacklist entry by domain', async function () {
@@ -31,7 +49,9 @@ describe('EmailDomainBlacklistService', function () {
 
     describe('updateDomainBlacklistEntry', function () {
         before(async function() {
-            await deleteAllRows();
+            if (!USE_DYNAMODB) {
+                await deleteAllRows();
+            }
         });
 
         it('creates the blacklist entry when one does not already exist', async function () {
@@ -45,7 +65,11 @@ describe('EmailDomainBlacklistService', function () {
 
         it('updates the existing blacklist entry', async function () {
             let blacklistEntry = await EmailDomainBlacklistService().getDomainBlacklistEntry(DOMAIN_NAME);
-            expect(blacklistEntry).property('isTemporaryEmailService', null);
+            if (USE_DYNAMODB) {
+                expect(blacklistEntry).not.have.property('isTemporaryEmailService');
+            } else {
+                expect(blacklistEntry).property('isTemporaryEmailService', null);
+            }
 
             await EmailDomainBlacklistService().updateDomainBlacklistEntry({
                 domainName: DOMAIN_NAME,
